Refactor some data classes out.
