+++ src/main/java/net/javasauce/ss/util/VersionRequest.java
@@ -1,0 +1,9 @@
+package net.javasauce.ss.util;
+
+/**
+ * Created by covers1624 on 7/22/25.
+ */
+public record VersionRequest(
+        String id,
+        String commitName
+) { }
